refactor(bot-manager): move trade processors into folder
